The logic for today boughts has been rewritten and a today boughts section has been added.
