--- src/components/HomeComponents/TodayBoughts/style.module.scss
@@ -1,47 +1,0 @@
-@use "../../../assets/scss/common/mixins" as mixin;
-
-.boughtsContainer {
-  @include mixin.flex(null, null, column, null);
-}
-.boughtsItemContainer {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-}
-.boughtsItem {
-  @include mixin.flex(null, null, column, null);
-}
-.boughtsItemImg {
-  border-radius: 5px;
-}
-.boughtsItemPrice {
-  @include mixin.font(16px, 700, 24px, #424242);
-}
-.boughtsItemName {
-  @include mixin.font(14px, 400, 16.8px, #62656a);
-  max-width: 187.4px;
-}
-.boughtsItemDelivery {
-  margin-top: 3px;
-  @include mixin.flex(center, null, null, 3px);
-  max-width: 240px;
-}
-.boughtsItemDeliveryTitle {
-  @include mixin.font(12px, 400, 14.4px, #0c54a1);
-  max-width: 160px;
-}
-.boughtsItemBtns {
-  margin-top: 20px;
-  @include mixin.flex(center, null, null, 10px);
-}
-.boughtsItemBtnsAddBucket {
-  @include mixin.button(
-    20px,
-    null,
-    null,
-    5px,
-    linear-gradient(88.03deg, #112b66 1.19%, #991a4e 97.58%),
-    null,
-    null,
-    #fff
-  );
-}
